backend/user-service/requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
Flask-SQLAlchemy==3.1.1
Flask-Cors==4.0.0
PyJWT==2.8.0
mysqlclient==2.2.0
jsonschema==4.19.0
python-dotenv==1.0.0
requests==2.31.0
PyMySQL==1.1.0 